Move context to the io module
