+++ src/backend/linux_raw/io/epoll.rs
@@ -71,7 +71,7 @@ use core::marker::PhantomData;
 use core::ptr::null;
 
 #[doc(inline)]
-pub use crate::utils::context::*;
+pub use crate::io::context::*;
 
 bitflags! {
     /// `EPOLL_*` for use with [`Epoll::new`].
